add interceptors, secured paths, jwt, user account registration

+++ Backend/src/main/java/com/example/demo/Repositories/UserRepository.java
@@ -3,8 +3,9 @@ package com.example.demo.Repositories;
 import com.example.demo.Entities.UserAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserAccount, Integer> {
+public interface UserRepository extends JpaRepository<UserAccount, Long> {
     UserAccount findByEmail(String email);
-    UserAccount findById(int id);
+    UserAccount findById(long id);
     UserAccount findByName(String username);
+
 }
